+++ mixed_congruential.cpp
@@ -1,7 +1,8 @@
 #include "linear_congruential.h"
 #include "mixed_congruential.h"
+#include <cstdint>
 
-MixedCongruentialGenerator::MixedCongruentialGenerator(long long modulus, long long multiplier, long long increment) :
+MixedCongruentialGenerator::MixedCongruentialGenerator(int64_t modulus, int64_t multiplier, int64_t increment) :
 	LinearCongruentialGenerator(modulus, 0), multiplier(multiplier), increment(increment) {};
 
 void MixedCongruentialGenerator::updateState() {
